Small refactoring to allow for more proofs (#20)
